requirements.txt
arize-phoenix>=11.8.0
python-dotenv>=1.0.0